利用elasticsearch完成文章（汉字/拼音）全文检索以及高亮显示

+++ monkey-article/src/main/java/com/monkey/monkeyarticle/pojo/ArticleComment.java
@@ -20,7 +20,7 @@ public class ArticleComment {
     private String content;
     private Long parentId;
     private Long replyId;
-    private Long likeSum;
+    private Integer likeSum;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Shanghai")
     private Date commentTime;
 }
